cat.yml
Hello Dev/Greg 

I have tried to make a small contribution with this translation in Catalan. A language understood by more than 10 million people in Europe (Catalonia, Valencia, Balearic Islands, southern France and Italian Sardinia)

The economy is above the European average and countries like Italy or Spain.
https://image.prntscr.com/image/sxfrEUmUQ2eBsurehWswhg.png

The language is in the top 20 of Wikipedia
http://royal.pingdom.com/wp-content/uploads/2012/06/wikipedia-top-20.001.jpg

And Barcelona is among the 10 cities in the world that more foreign investment receives
https://image.prntscr.com/image/_knn0W1eSRCeNug6odpGUA.png

I would love the Verge website to be available in catalan language! Thanks